DesktopLauncher: get the activated item from the filtered list.
When a more generic search has been narrowed down via a TreeModelFilter, the TreePath
that the IconView gives at the on_item_activated() callback is relative to the filtered
model, which may not be the same as the normal one. This means that the selected item
may not match what's in the full model at the given path, thus probably obtaining informations
of the wrong application.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>