DesktopLauncher: drop DesktopLauncher.searching global state variable.
On most machines its status is reverted to false before the majority of the items have determined their visibility.
This means that the item_matches_keyword() check is done regardless if the search is fresh or not.
Probably an inverse state variable (like "DesktopLauncher.refiltering") would work better, but for now just we are good
without a global state variable.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>