DesktopLauncher: made the page handling and search with the same start_text more robust.
The last-page property is now updated on-the-fly while processing item visibilty. This provides
seamless integration with both standard, GMenu-based search and with the subsearch using filters.

on_application_found() now directly appends the application to the ListStore with one instruction
(ListStore.insert_with_values()).
Previously the iter was obtained via append() and then a call to set() actually set the application
informations. This fired the row_changed() signal twice, thus doubling the amount of pages detected.

In some cases, an iter is processed twice when refiltering on a subsearch. This is workarounded by
saving the previous iter and comparing it with the new. The issue needs investigation.

Also, the global search state (removed in 952b9fb) has been reintroduced, as the blocking bug has been
fixed in 4464476.
This avoids redundant checks when an application has been added via on_application_found().

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>